refactor(ui): move clip split/join/region ops into arrangement domain
Tranche 3b-2: DuplicateNoteClip, SplitAudioClip, SplitNoteClip,
SplitSelectedAtPlayhead, JoinSelectedClips, DeleteClipsInRegion,
SplitClipsAtRegion, CreateClipFromSelection and
CreateNoteClipFromSelection now live in domains/arrangement.rs,
along with the join_audio_clips/join_note_clips helpers.

ArrangementCtx gains playhead_samples/playhead_beats;
ArrangementAction gains close_context_menu. Five new unit tests
cover split bounds, join validation/merging, and note clip
creation guards.
